Refactor all SecOps tools to support both MCP and Function Calling
- Extracted core logic from all MCP tools into shared `*_logic.py` modules.
- Updated existing MCP tools to use the shared logic.
- Created `*_function_calling.py` modules for all tool categories, providing standalone OpenAI Function Calling support.
- Updated `server/secops/README.md` and `docs/function_calling.md` to document the new capabilities.
