+++ server/secops/secops_mcp/tools/ioc_logic.py
@@ -1,0 +1,83 @@
+# Copyright 2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Shared logic for IoC matches tools."""
+
+import logging
+from datetime import datetime, timedelta, timezone
+from typing import Optional
+
+from secops_mcp.server import get_chronicle_client
+
+# Configure logging
+logger = logging.getLogger('secops-mcp')
+
+def get_ioc_matches_impl(
+    project_id: Optional[str] = None,
+    customer_id: Optional[str] = None,
+    hours_back: int = 24,
+    max_matches: int = 20,
+    region: Optional[str] = None,
+) -> str:
+    """Core implementation for getting IoC matches."""
+    try:
+        chronicle = get_chronicle_client(project_id, customer_id, region)
+
+        end_time = datetime.now(timezone.utc)
+        start_time = end_time - timedelta(hours=hours_back)
+
+        iocs = chronicle.list_iocs(
+            start_time=start_time, end_time=end_time, max_matches=max_matches
+        )
+
+        # Handle different possible response formats
+        matches = []
+        if isinstance(iocs, dict) and 'matches' in iocs:
+            matches = iocs.get('matches', [])
+        elif isinstance(iocs, list):
+            matches = iocs
+
+        if not matches:
+            return 'No IoC matches found for the specified time range.'
+
+        result = f'Found {len(matches)} IoC matches:\n\n'
+
+        for i, match in enumerate(matches, 1):
+            # Get the indicator information
+            indicator_type = 'Unknown'
+            indicator_value = 'Unknown'
+            sources = []
+
+            # Try to extract artifactIndicator differently based on response format
+            if isinstance(match, dict):
+                if 'artifactIndicator' in match and isinstance(
+                    match['artifactIndicator'], dict
+                ):
+                    # Get the first key-value pair from artifactIndicator
+                    indicator_dict = match.get('artifactIndicator', {})
+                    if indicator_dict:
+                        indicator_type = next(iter(indicator_dict.keys()), 'Unknown')
+                        indicator_value = next(iter(indicator_dict.values()), 'Unknown')
+
+                sources = match.get('sources', [])
+
+            sources_str = ', '.join(sources) if sources else 'Unknown'
+
+            result += f'IoC {i}:\n'
+            result += f'Type: {indicator_type}\n'
+            result += f'Value: {indicator_value}\n'
+            result += f'Sources: {sources_str}\n\n'
+
+        return result
+    except Exception as e:
+        return f'Error retrieving IoC matches: {str(e)}'
