+++ server/secops/secops_mcp/tools/parser_management.py
@@ -13,12 +13,17 @@
 # limitations under the License.
 """Security Operations MCP tools for parser management."""
 
-import base64
-import json
 import logging
 from typing import Any, Dict, List, Optional
 
-from secops_mcp.server import get_chronicle_client, server
+from secops_mcp.server import server
+from secops_mcp.tools.parser_logic import (
+    activate_parser_impl,
+    create_parser_impl,
+    deactivate_parser_impl,
+    get_parser_impl,
+    run_parser_against_sample_logs_impl,
+)
 
 
 # Configure logging
@@ -111,36 +116,14 @@ async def create_parser(
         - Monitor parsing success and adjust the parser configuration if needed.
         - Create detection rules that leverage the parsed UDM fields.
     """
-    try:
-        logger.info(f'Creating parser for log type: {log_type}')
-
-        
-        
-        chronicle = get_chronicle_client(project_id, customer_id, region)
-
-        # Create the parser
-        parser = chronicle.create_parser(
-            log_type=log_type,
-            parser_code=parser_code,
-            validated_on_empty_logs=validated_on_empty_logs
-        )
-
-        # Extract parser ID from the response
-        parser_id = parser.get("name", "").split("/")[-1]
-        state = parser.get("state", "Unknown")
-        
-        result = f'Successfully created parser for log type: {log_type}\n'
-        result += f'Parser ID: {parser_id}\n'
-        result += f'State: {state}\n'
-        
-        if validated_on_empty_logs:
-            result += 'Parser was validated on empty logs during creation.'
-        
-        return result
-
-    except Exception as e:
-        logger.error(f'Error creating parser for log type {log_type}: {str(e)}', exc_info=True)
-        return f'Error creating parser for log type {log_type}: {str(e)}'
+    return create_parser_impl(
+        log_type=log_type,
+        parser_code=parser_code,
+        project_id=project_id,
+        customer_id=customer_id,
+        region=region,
+        validated_on_empty_logs=validated_on_empty_logs,
+    )
 
 @server.tool()
 async def get_parser(
@@ -197,46 +180,13 @@ async def get_parser(
         - Use the configuration as a template for creating parsers for similar log types.
         - Activate or deactivate the parser based on your requirements.
     """
-    try:
-        logger.info(f'Getting parser {parser_id} for log type: {log_type}')
-
-        
-        
-        chronicle = get_chronicle_client(project_id, customer_id, region)
-
-        # Get the parser
-        parser = chronicle.get_parser(log_type=log_type, id=parser_id)
-
-        parser_name = parser.get("name", "").split("/")[-1]
-        state = parser.get("state", "Unknown")
-        create_time = parser.get("createTime", "Unknown")
-        
-        result = f'Parser Details:\n\n'
-        result += f'Parser ID: {parser_name}\n'
-        result += f'Log Type: {log_type}\n'
-        result += f'State: {state}\n'
-        result += f'Created: {create_time}\n\n'
-        
-        # Extract and decode parser code if available
-        parser_code = parser.get("text", "")
-        if not parser_code and "cbn" in parser:
-            # Decode base64 encoded parser code
-            try:
-                parser_code = base64.b64decode(parser["cbn"]).decode('utf-8')
-            except Exception as decode_error:
-                logger.warning(f"Failed to decode parser code: {decode_error}")
-                parser_code = "Could not decode parser code"
-        
-        if parser_code:
-            result += f'Parser Code:\n{parser_code}\n'
-        else:
-            result += 'Parser code not available in response.\n'
-        
-        return result
-
-    except Exception as e:
-        logger.error(f'Error getting parser {parser_id} for log type {log_type}: {str(e)}', exc_info=True)
-        return f'Error getting parser {parser_id} for log type {log_type}: {str(e)}'
+    return get_parser_impl(
+        log_type=log_type,
+        parser_id=parser_id,
+        project_id=project_id,
+        customer_id=customer_id,
+        region=region,
+    )
 
 @server.tool()
 async def activate_parser(
@@ -293,25 +243,13 @@ async def activate_parser(
         - Create detection rules that leverage the newly parsed UDM fields.
         - Set up monitoring for the log type to ensure continued parsing success.
     """
-    try:
-        logger.info(f'Activating parser {parser_id} for log type: {log_type}')
-
-        
-        
-        chronicle = get_chronicle_client(project_id, customer_id, region)
-
-        # Activate the parser
-        chronicle.activate_parser(log_type=log_type, id=parser_id)
-
-        result = f'Successfully activated parser for log type: {log_type}\n'
-        result += f'Parser ID: {parser_id}\n'
-        result += 'The parser is now active and will process incoming logs of this type.'
-        
-        return result
-
-    except Exception as e:
-        logger.error(f'Error activating parser {parser_id} for log type {log_type}: {str(e)}', exc_info=True)
-        return f'Error activating parser {parser_id} for log type {log_type}: {str(e)}'
+    return activate_parser_impl(
+        log_type=log_type,
+        parser_id=parser_id,
+        project_id=project_id,
+        customer_id=customer_id,
+        region=region,
+    )
 
 @server.tool()
 async def deactivate_parser(
@@ -372,25 +310,13 @@ async def deactivate_parser(
         - Test and validate any replacement parser before activating it.
         - Document the reason for deactivation for operational tracking.
     """
-    try:
-        logger.info(f'Deactivating parser {parser_id} for log type: {log_type}')
-
-        
-        
-        chronicle = get_chronicle_client(project_id, customer_id, region)
-
-        # Deactivate the parser
-        chronicle.deactivate_parser(log_type=log_type, id=parser_id)
-
-        result = f'Successfully deactivated parser for log type: {log_type}\n'
-        result += f'Parser ID: {parser_id}\n'
-        result += 'WARNING: Incoming logs of this type will not be parsed until a parser is activated.'
-        
-        return result
-
-    except Exception as e:
-        logger.error(f'Error deactivating parser {parser_id} for log type {log_type}: {str(e)}', exc_info=True)
-        return f'Error deactivating parser {parser_id} for log type {log_type}: {str(e)}'
+    return deactivate_parser_impl(
+        log_type=log_type,
+        parser_id=parser_id,
+        project_id=project_id,
+        customer_id=customer_id,
+        region=region,
+    )
 
 @server.tool()
 async def run_parser_against_sample_logs(
@@ -496,76 +422,13 @@ async def run_parser_against_sample_logs(
         - Activate the parser using `activate_parser` to put it into production.
         - Ingest real logs using `ingest_raw_log` and verify parsing works in production.
     """
-    try:
-        logger.info(f'Running parser test for log type: {log_type} with {len(sample_logs)} sample logs')
-
-        # Validate input constraints
-        if len(sample_logs) > 1000:
-            return "Error: Maximum of 1000 sample logs allowed per test."
-        
-        total_size = sum(len(log.encode('utf-8')) for log in sample_logs)
-        if total_size > 50 * 1024 * 1024:  # 50MB
-            return "Error: Total sample logs size exceeds 50MB limit."
-        
-        for i, log in enumerate(sample_logs):
-            if len(log.encode('utf-8')) > 10 * 1024 * 1024:  # 10MB
-                return f"Error: Sample log {i+1} exceeds 10MB size limit."
-
-        
-        
-        chronicle = get_chronicle_client(project_id, customer_id, region)
-
-        # Run the parser
-        result = chronicle.run_parser(
-            log_type=log_type,
-            parser_code=parser_code,
-            parser_extension_code=parser_extension_code,
-            logs=sample_logs,
-            statedump_allowed=statedump_allowed
-        )
-
-        # Process and format the results
-        response = f'Parser test results for log type: {log_type}\n'
-        response += f'Tested {len(sample_logs)} sample log(s)\n\n'
-
-        if "runParserResults" in result:
-            for i, parser_result in enumerate(result["runParserResults"]):
-                response += f'Log {i+1} Results:\n'
-                
-                # Check for parsed events
-                if "parsedEvents" in parser_result and parser_result["parsedEvents"]:
-                    parsed_events = parser_result["parsedEvents"]
-                    if isinstance(parsed_events, dict) and "events" in parsed_events:
-                        events = parsed_events["events"]
-                        response += f'  Successfully parsed {len(events)} UDM event(s)\n'
-                        
-                        # Show first event details
-                        if events:
-                            first_event = events[0]
-                            if "event" in first_event:
-                                event_data = first_event["event"]
-                                if "metadata" in event_data:
-                                    metadata = event_data["metadata"]
-                                    event_type = metadata.get("eventType", "Unknown")
-                                    response += f'  Event Type: {event_type}\n'
-                                    if "description" in metadata:
-                                        response += f'  Description: {metadata["description"]}\n'
-                    else:
-                        response += f'  Parsed events: {parsed_events}\n'
-                else:
-                    response += '  No parsed events generated\n'
-                
-                # Check for errors
-                if "errors" in parser_result and parser_result["errors"]:
-                    errors = parser_result["errors"]
-                    response += f'  Parsing errors: {errors}\n'
-                
-                response += '\n'
-        else:
-            response += f'Unexpected result format: {result}'
-
-        return response
-
-    except Exception as e:
-        logger.error(f'Error running parser test for log type {log_type}: {str(e)}', exc_info=True)
-        return f'Error running parser test for log type {log_type}: {str(e)}' 
+    return run_parser_against_sample_logs_impl(
+        log_type=log_type,
+        parser_code=parser_code,
+        sample_logs=sample_logs,
+        project_id=project_id,
+        customer_id=customer_id,
+        region=region,
+        parser_extension_code=parser_extension_code,
+        statedump_allowed=statedump_allowed,
+    )
