+++ server/secops/secops_mcp/tools/rules_function_calling.py
@@ -1,0 +1,370 @@
+# Copyright 2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+This module provides a standalone implementation of the security rules functionality
+using the OpenAI Function Calling format.
+"""
+
+import logging
+from typing import Any, Dict
+
+from secops_mcp.tools.rules_logic import (
+    create_rule_impl,
+    get_detection_rule_impl,
+    get_rule_detections_impl,
+    list_rule_errors_impl,
+    list_rules_impl,
+    search_security_rules_impl,
+    test_rule_impl,
+    validate_rule_impl,
+)
+
+# Configure logging
+logger = logging.getLogger("secops-mcp-fc")
+
+
+# 1. The Function Definitions (JSON Schema)
+
+LIST_RULES_SCHEMA = {
+    "type": "function",
+    "name": "list_security_rules",
+    "description": "List security detection rules configured in Chronicle SIEM, with support for pagination.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+            "page_size": {
+                "type": "integer",
+                "description": "Maximum number of rules to return. Defaults to 100. Max is 1000.",
+                "default": 100,
+            },
+            "page_token": {
+                "type": "string",
+                "description": "Page token for pagination.",
+            },
+        },
+        "required": [],
+        "additionalProperties": False,
+    },
+}
+
+SEARCH_RULES_SCHEMA = {
+    "type": "function",
+    "name": "search_security_rules",
+    "description": "Search security detection rules configured in Chronicle SIEM using a regex pattern.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "query": {
+                "type": "string",
+                "description": "Regex string to use for searching SecOps rules.",
+            },
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+        },
+        "required": ["query"],
+        "additionalProperties": False,
+    },
+}
+
+GET_RULE_SCHEMA = {
+    "type": "function",
+    "name": "get_detection_rule",
+    "description": "Retrieve the complete definition and metadata of a specific detection rule from Chronicle SIEM.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "rule_id": {
+                "type": "string",
+                "description": "Unique ID of the detection rule to retrieve.",
+            },
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+        },
+        "required": ["rule_id"],
+        "additionalProperties": False,
+    },
+}
+
+GET_RULE_DETECTIONS_SCHEMA = {
+    "type": "function",
+    "name": "get_rule_detections",
+    "description": "Retrieves historical detections generated by a specific Chronicle SIEM rule.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "rule_id": {
+                "type": "string",
+                "description": "Unique ID of the rule to list detections for.",
+            },
+            "alert_state": {
+                "type": "string",
+                "description": "Filter by alert state. Valid values: 'UNSPECIFIED', 'NOT_ALERTING', 'ALERTING'.",
+                "enum": ["UNSPECIFIED", "NOT_ALERTING", "ALERTING"],
+            },
+            "page_size": {
+                "type": "integer",
+                "description": "Maximum number of detections to return in a single response.",
+            },
+            "page_token": {
+                "type": "string",
+                "description": "Token to retrieve the next page of results for pagination.",
+            },
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+        },
+        "required": ["rule_id"],
+        "additionalProperties": False,
+    },
+}
+
+LIST_RULE_ERRORS_SCHEMA = {
+    "type": "function",
+    "name": "list_rule_errors",
+    "description": "Lists execution errors for a specific Chronicle SIEM rule.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "rule_id": {
+                "type": "string",
+                "description": "Unique ID of the rule to list errors for.",
+            },
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+        },
+        "required": ["rule_id"],
+        "additionalProperties": False,
+    },
+}
+
+CREATE_RULE_SCHEMA = {
+    "type": "function",
+    "name": "create_rule",
+    "description": "Create a new detection rule in Chronicle SIEM.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "rule_text": {
+                "type": "string",
+                "description": "Complete YARA-L 2.0 rule definition including rule metadata, events, and conditions.",
+            },
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+        },
+        "required": ["rule_text"],
+        "additionalProperties": False,
+    },
+}
+
+TEST_RULE_SCHEMA = {
+    "type": "function",
+    "name": "test_rule",
+    "description": "Test a detection rule against historical data in Chronicle SIEM.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "rule_text": {
+                "type": "string",
+                "description": "Complete YARA-L 2.0 rule definition to test.",
+            },
+            "hours_back": {
+                "type": "integer",
+                "description": "How many hours of historical data to test against. Defaults to 168 (7 days).",
+                "default": 168,
+            },
+            "max_results": {
+                "type": "integer",
+                "description": "Maximum number of detection results to return. Defaults to 100.",
+                "default": 100,
+            },
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+        },
+        "required": ["rule_text"],
+        "additionalProperties": False,
+    },
+}
+
+VALIDATE_RULE_SCHEMA = {
+    "type": "function",
+    "name": "validate_rule",
+    "description": "Validate a YARA-L 2.0 detection rule syntax in Chronicle SIEM.",
+    "parameters": {
+        "type": "object",
+        "properties": {
+            "rule_text": {
+                "type": "string",
+                "description": "Complete YARA-L 2.0 rule definition to validate.",
+            },
+            "project_id": {
+                "type": "string",
+                "description": "Google Cloud project ID. Optional if configured in environment.",
+            },
+            "customer_id": {
+                "type": "string",
+                "description": "Chronicle customer ID. Optional if configured in environment.",
+            },
+            "region": {
+                "type": "string",
+                "description": "Chronicle region (e.g., 'us', 'europe'). Optional if configured in environment.",
+            },
+        },
+        "required": ["rule_text"],
+        "additionalProperties": False,
+    },
+}
+
+
+# 2. The Tool Execution Logic
+
+async def execute_list_rules_tool(arguments: Dict[str, Any]) -> Dict[str, Any]:
+    return list_rules_impl(
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+        page_size=arguments.get("page_size", 100),
+        page_token=arguments.get("page_token"),
+    )
+
+async def execute_search_rules_tool(arguments: Dict[str, Any]) -> Dict[str, Any]:
+    return search_security_rules_impl(
+        query=arguments.get("query"),
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+    )
+
+async def execute_get_rule_tool(arguments: Dict[str, Any]) -> Dict[str, Any]:
+    return get_detection_rule_impl(
+        rule_id=arguments.get("rule_id"),
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+    )
+
+async def execute_get_rule_detections_tool(arguments: Dict[str, Any]) -> Dict[str, Any]:
+    return get_rule_detections_impl(
+        rule_id=arguments.get("rule_id"),
+        alert_state=arguments.get("alert_state"),
+        page_size=arguments.get("page_size"),
+        page_token=arguments.get("page_token"),
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+    )
+
+async def execute_list_rule_errors_tool(arguments: Dict[str, Any]) -> Dict[str, Any]:
+    return list_rule_errors_impl(
+        rule_id=arguments.get("rule_id"),
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+    )
+
+async def execute_create_rule_tool(arguments: Dict[str, Any]) -> str:
+    return create_rule_impl(
+        rule_text=arguments.get("rule_text"),
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+    )
+
+async def execute_test_rule_tool(arguments: Dict[str, Any]) -> str:
+    return test_rule_impl(
+        rule_text=arguments.get("rule_text"),
+        hours_back=arguments.get("hours_back", 168),
+        max_results=arguments.get("max_results", 100),
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+    )
+
+async def execute_validate_rule_tool(arguments: Dict[str, Any]) -> str:
+    return validate_rule_impl(
+        rule_text=arguments.get("rule_text"),
+        project_id=arguments.get("project_id"),
+        customer_id=arguments.get("customer_id"),
+        region=arguments.get("region"),
+    )
